project_creation: Project -> BaseProject
Name and module are more suitable
