+++ tests/commands/project_creation/test_linux_project.py
@@ -25,8 +25,8 @@ import os
 from tempfile import gettempdir
 from unittest import TestCase
 
-from s2e_env.commands.project_creation.linux_project import LinuxProject
-from s2e_env.commands.project_creation.target import Target
+from s2e_env.commands.project_creation import LinuxProject
+from s2e_env.commands.project_creation import Target
 
 from . import DATA_DIR, monkey_patch_project
