+++ tests/commands/project_creation/test_target.py
@@ -24,11 +24,11 @@ SOFTWARE.
 import os
 from unittest import TestCase
 
-from s2e_env.commands.project_creation.cgc_project import CGCProject
-from s2e_env.commands.project_creation.linux_project import LinuxProject
-from s2e_env.commands.project_creation.windows_project import WindowsProject, \
+from s2e_env.commands.project_creation import CGCProject
+from s2e_env.commands.project_creation import LinuxProject
+from s2e_env.commands.project_creation import WindowsProject, \
         WindowsDLLProject, WindowsDriverProject
-from s2e_env.commands.project_creation.target import Target
+from s2e_env.commands.project_creation import Target
 
 from . import DATA_DIR
